refactor: Move two output functions to the output package
