+++ cmd/oras/internal/display/metadata/template/manifest_fetch.go
@@ -22,6 +22,7 @@ import (
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 	"oras.land/oras/cmd/oras/internal/display/metadata"
 	"oras.land/oras/cmd/oras/internal/display/metadata/model"
+	"oras.land/oras/cmd/oras/internal/output"
 )
 
 // manifestFetchHandler handles JSON metadata output for manifest fetch events.
@@ -44,5 +45,5 @@ func (h *manifestFetchHandler) OnFetched(path string, desc ocispec.Descriptor, c
 	if err := json.Unmarshal(content, &manifest); err != nil {
 		manifest = nil
 	}
-	return parseAndWrite(h.out, model.NewFetched(path, desc, manifest), h.template)
+	return output.ParseAndWrite(h.out, model.NewFetched(path, desc, manifest), h.template)
 }
