+++ cmd/oras/internal/output/template.go
@@ -13,24 +13,23 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package template
+package output
 
 import (
 	"io"
 	"text/template"
 
 	"github.com/Masterminds/sprig/v3"
-	"oras.land/oras/cmd/oras/internal/display/utils"
 )
 
-func parseAndWrite(out io.Writer, object any, templateStr string) error {
+func ParseAndWrite(out io.Writer, object any, templateStr string) error {
 	// parse template
 	t, err := template.New("format output").Funcs(sprig.FuncMap()).Parse(templateStr)
 	if err != nil {
 		return err
 	}
 	// convert object to map[string]any
-	converted, err := utils.ToMap(object)
+	converted, err := ToMap(object)
 	if err != nil {
 		return err
 	}
